Refactor: put game loop to a separate file
